Update main function
