+++ src/Core/core.scala
@@ -188,25 +188,3 @@ class Core(confIn: FlexpretConfiguration) extends Module
   //io.int_exts <> datapath.io.int_exts
 
 }
-
-object CoreMain {
-  def main(args: Array[String]): Unit = { 
-    
-    val confString = args(0)
-    val chiselArgs = args.slice(1, args.length)
-
-    val parsed = """(\d+)t(.*)-(\d+)i-(\d+)d.*-(.*)""".r.findFirstMatchIn(confString)
-    val coreConfig = new FlexpretConfiguration(
-      parsed.get.group(1).toInt,
-      !parsed.get.group(2).isEmpty,
-      parsed.get.group(3).toInt,
-      parsed.get.group(4).toInt,
-      confString contains "mul",
-      parsed.get.group(5)
-      )
-      
-    // Pass configuration to FlexPRET processor.
-    chiselMain( chiselArgs, () => Module(new Core(coreConfig)) )
-
-   }
-}
